Move unit tests to dedicated services
